% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.2all.R
\name{odds.dec2all}
\alias{odds.dec2all}
\title{Convert Decimal Odds to all other formats}
\usage{
odds.dec2all(x)
}
\arguments{
\item{x}{A vector of Decimal odds}
}
\value{
A dataframe
}
\description{
Convert Decimal Odds to all other formats
}
\examples{
odds.dec2all(c(1.93,2.05))
}

