% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.dec2hk.R
\name{odds.dec2hk}
\alias{odds.dec2hk}
\title{Convert Decimal Odds to Hong Kong odds}
\usage{
odds.dec2hk(x)
}
\arguments{
\item{x}{A vector of Decimal odds}
}
\value{
A vector of Hong Kong odds
}
\description{
Convert Decimal Odds to Hong Kong odds
}
\examples{
odds.dec2hk(c(1.93,2.05))
}

