% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.dec2indo.R
\name{odds.dec2indo}
\alias{odds.dec2indo}
\title{Convert Decimal Odds to Indonesian odds}
\usage{
odds.dec2indo(x)
}
\arguments{
\item{x}{A vector of Decimal odds}
}
\value{
A vector of Indonesian odds
}
\description{
Convert Decimal Odds to Indonesian odds
}
\examples{
odds.dec2indo(c(1.93,2.05))
}
