% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.dec2malay.R
\name{odds.dec2malay}
\alias{odds.dec2malay}
\title{Convert Decimal Odds to Malaysian odds}
\usage{
odds.dec2malay(x)
}
\arguments{
\item{x}{A vector of Decimal odds}
}
\value{
A vector of Malaysian odds
}
\description{
Convert Decimal Odds to Malaysian odds
}
\examples{
odds.dec2malay(c(1.93,2.05))
}
