% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.dec2prob.R
\name{odds.dec2prob}
\alias{odds.dec2prob}
\title{Convert Decimal Odds to Probabilities}
\usage{
odds.dec2prob(x)
}
\arguments{
\item{x}{A vector of Decimal odds}
}
\value{
A vector of Probabilities
}
\description{
Convert Decimal Odds to Probabilities
}
\examples{
odds.dec2prob(c(1.93,2.05))
}
