% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/odds.dec2us.R
\name{odds.dec2us}
\alias{odds.dec2us}
\title{Convert Decimal Odds to US Odds}
\usage{
odds.dec2us(x)
}
\arguments{
\item{x}{A vector of Decimal odds}
}
\value{
A vector of US odds
}
\description{
Convert Decimal Odds to US Odds
}
\examples{
odds.dec2us(c(1.93,2.05))
}

