% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/odds.fv.R
\name{odds.fv}
\alias{odds.fv}
\title{Calculate the fair ( vigorish free) odds for a vector of vigged odds.}
\usage{
odds.fv(a, b, ..., input = "us", output = "us")
}
\arguments{
\item{a}{Odds of the format defined in input}

\item{b}{Odds of the format defined in input}

\item{...}{Odds of the format defined in input}

\item{input}{Default to US Odds. Valid selections are  "us" for US odds,
"prob" for Probability ,"dec" for Decimal odds,"malay" for Malaysian odds,
"hk" for Hong Kong odds, "indo" for Indonesian odds}

\item{output}{Default to US Odds. Valid selections are  "us" for US odds,
"prob" for Probability ,"dec" for Decimal odds,"malay" for Malaysian odds,
"hk" for Hong Kong odds, "indo" for Indonesian odds}
}
\value{
A vector of the fair value of the the given vigged values in the format chosen in output
}
\description{
Calculate the fair ( vigorish free) odds for a vector of vigged odds.
}
\examples{
odds.fv(-120,110)
odds.fv(2.05,3.12,2.90,input="dec",output="prob")
}

