% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.2all.R
\name{odds.hk2all}
\alias{odds.hk2all}
\title{Convert Hong Kong Odds to all other formats}
\usage{
odds.hk2all(x)
}
\arguments{
\item{x}{A vector of Hong Kong odds}
}
\value{
A dataframe
}
\description{
Convert Hong Kong Odds to all other formats
}
\examples{
odds.hk2all(c(1.93,0.05))
}

