% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/odds.hk2dec.R
\name{odds.hk2dec}
\alias{odds.hk2dec}
\title{Convert Hong Kong Odds to Decimal Odds}
\usage{
odds.hk2dec(x)
}
\arguments{
\item{x}{A vector of Hong Kong odds}
}
\value{
A vector of Decimal odds
}
\description{
Convert Hong Kong Odds to Decimal Odds
}
\examples{
odds.hk2dec(c(1.93,0.05))
}

