% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/odds.hk2indo.R
\name{odds.hk2indo}
\alias{odds.hk2indo}
\title{Convert Hong Kong Odds to Indonesian Odds}
\usage{
odds.hk2indo(x)
}
\arguments{
\item{x}{A vector of Hong Kong odds}
}
\value{
A vector of Indonesian odds
}
\description{
Convert Hong Kong Odds to Indonesian Odds
}
\examples{
odds.hk2indo(c(1.93,0.05))
}

