% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.hk2malay.R
\name{odds.hk2malay}
\alias{odds.hk2malay}
\title{Convert Hong Kong Odds to Malaysian Odds}
\usage{
odds.hk2malay(x)
}
\arguments{
\item{x}{A vector of Hong Kong odds}
}
\value{
A vector of Malaysian odds
}
\description{
Convert Hong Kong Odds to Malaysian Odds
}
\examples{
odds.hk2malay(c(1.93,0.05))
}
