% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.hk2prob.R
\name{odds.hk2prob}
\alias{odds.hk2prob}
\title{Convert Hong Kong Odds to Probabilities}
\usage{
odds.hk2prob(x)
}
\arguments{
\item{x}{A vector of Hong Kong odds}
}
\value{
A vector of Probabilities
}
\description{
Convert Hong Kong Odds to Probabilities
}
\examples{
odds.hk2us(c(1.93,0.05))
}
