% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/odds.hk2us.R
\name{odds.hk2us}
\alias{odds.hk2us}
\title{Convert Hong Kong Odds to US Odds}
\usage{
odds.hk2us(x)
}
\arguments{
\item{x}{A vector of Hong Kong odds}
}
\value{
A vector of US odds
}
\description{
Convert Hong Kong Odds to US Odds
}
\examples{
odds.hk2us(c(1.93,0.05))
}

