% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.2all.R
\name{odds.indo2all}
\alias{odds.indo2all}
\title{Convert Indonesian to all other formats}
\usage{
odds.indo2all(x)
}
\arguments{
\item{x}{A vector of Indonesian odds}
}
\value{
A vector of US odds
}
\description{
Convert Indonesian to all other formats
}
\examples{
odds.indo2all(c(1.93,2.05))
}

