% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.indo2dec.R
\name{odds.indo2dec}
\alias{odds.indo2dec}
\title{Convert Indonesian odds to Decimal odds}
\usage{
odds.indo2dec(x)
}
\arguments{
\item{x}{A vector of Indonesian odds}
}
\value{
A vector of Hong Kong odds
}
\description{
Convert Indonesian odds to Decimal odds
}
\examples{
odds.indo2dec(c(1.93,2.05))
}

