% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.indo2hk.R
\name{odds.indo2hk}
\alias{odds.indo2hk}
\title{Convert Indonesian odds to Hong Kong odds}
\usage{
odds.indo2hk(x)
}
\arguments{
\item{x}{A vector of Indonesian odds}
}
\value{
A vector of Hong Kong odds
}
\description{
Convert Indonesian odds to Hong Kong odds
}
\examples{
odds.indo2hk(c(1.93,2.05))
}

