% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.indo2malay.R
\name{odds.indo2malay}
\alias{odds.indo2malay}
\title{Convert Indonesian odds to Malaysian odds}
\usage{
odds.indo2malay(x)
}
\arguments{
\item{x}{A vector of Indonesian odds}
}
\value{
A vector of Malaysian odds
}
\description{
Convert Indonesian odds to Malaysian odds
}
\examples{
odds.indo2malay(c(1.93,2.05))
}

