% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/odds.indo2prob.R
\name{odds.indo2prob}
\alias{odds.indo2prob}
\title{Convert Indonesian odds to Probabilities}
\usage{
odds.indo2prob(x)
}
\arguments{
\item{x}{A vector of Indonesian odds}
}
\value{
A vector of Probabilities
}
\description{
Convert Indonesian odds to Probabilities
}
\examples{
odds.indo2hk(c(1.93,2.05))
}

