% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/odds.indo2us.R
\name{odds.indo2us}
\alias{odds.indo2us}
\title{Convert Indonesian odds to US odds}
\usage{
odds.indo2us(x)
}
\arguments{
\item{x}{A vector of Indonesian odds}
}
\value{
A vector of US odds
}
\description{
Convert Indonesian odds to US odds
}
\examples{
odds.indo2us(c(1.93,2.05))
}

