% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.2all.R
\name{odds.malay2all}
\alias{odds.malay2all}
\title{Convert Malaysian Odds to all other formats}
\usage{
odds.malay2all(x)
}
\arguments{
\item{x}{A vector of Malaysian odds}
}
\value{
A data.frame
}
\description{
Convert Malaysian Odds to all other formats
}
\examples{
odds.malay2all(c(0.5,-0.6))
}
