% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.malay2dec.R
\name{odds.malay2dec}
\alias{odds.malay2dec}
\title{Convert Malaysian odds to Decimal odds}
\usage{
odds.malay2dec(x)
}
\arguments{
\item{x}{A vector of Malaysian odds}
}
\value{
A vector of Decimal odds
}
\description{
Convert Malaysian odds to Decimal odds
}
\examples{
odds.malay2dec(c(0.5,-0.6))
}

