% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.malay2hk.R
\name{odds.malay2hk}
\alias{odds.malay2hk}
\title{Convert Malaysian odds to Hong Kong odds}
\usage{
odds.malay2hk(x)
}
\arguments{
\item{x}{A vector of Malaysian odds}
}
\value{
A vector of Hong Kong odds
}
\description{
Convert Malaysian odds to Hong Kong odds
}
\examples{
odds.malay2hk(c(1.93,2.05))
}

