% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/odds.malay2indo.R
\name{odds.malay2indo}
\alias{odds.malay2indo}
\title{Convert Malaysian odds to Indonesian odds}
\usage{
odds.malay2indo(x)
}
\arguments{
\item{x}{A vector of Malaysian odds}
}
\value{
A vector of Indonesian odds
}
\description{
Convert Malaysian odds to Indonesian odds
}
\examples{
odds.malay2indo(c(1.93,2.05))
}

