% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.malay2prob.R
\name{odds.malay2prob}
\alias{odds.malay2prob}
\title{Convert Malaysian odds to Probabilities}
\usage{
odds.malay2prob(x)
}
\arguments{
\item{x}{A vector of Malaysian odds}
}
\value{
A vector of Probabilities
}
\description{
Convert Malaysian odds to Probabilities
}
\examples{
odds.malay2prob(c(1.93,2.05))
}
