% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.malay2us.R
\name{odds.malay2us}
\alias{odds.malay2us}
\title{Convert Malaysian odds to US odds}
\usage{
odds.malay2us(x)
}
\arguments{
\item{x}{A vector of Malaysian odds}
}
\value{
A vector of US odds
}
\description{
Convert Malaysian odds to US odds
}
\examples{
odds.malay2us(c(0.5,-0.6))
}

