% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.parlay.R
\name{odds.parlay}
\alias{odds.parlay}
\title{Transfer a vector to odds into correct parlay odds}
\usage{
odds.parlay(..., input = "us", output = "us")
}
\arguments{
\item{...}{Odds of the format defined in input}

\item{input}{Default to US Odds. Valid selections are  "us" for US odds,
"prob" for Probability ,"dec" for Decimal odds,"malay" for Malaysian odds,
"hk" for Hong Kong odds, "indo" for Indonesian odds}

\item{output}{Default to US Odds. Valid selections are  "us" for US odds,
"prob" for Probability ,"dec" for Decimal odds,"malay" for Malaysian odds,
"hk" for Hong Kong odds, "indo" for Indonesian odds}
}
\value{
Parlay odds in desired odds format
}
\description{
Transfer a vector to odds into correct parlay odds
}
\examples{
odds.parlay(c(-105,-110),output="dec")
}
