% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/odds.2all.R
\name{odds.prob2all}
\alias{odds.prob2all}
\title{Convert Probabilities to  to all other formats}
\usage{
odds.prob2all(x)
}
\arguments{
\item{x}{A vector of Probabilities}
}
\value{
A dataframe
}
\description{
Convert Probabilities to  to all other formats
}
\examples{
odds.prob2all(c(0.5,0.6))
}

