% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.prob2dec.R
\name{odds.prob2dec}
\alias{odds.prob2dec}
\title{Convert Probabilities to Decimal odds}
\usage{
odds.prob2dec(x)
}
\arguments{
\item{x}{A vector of Probabilities}
}
\value{
A vector of Decimal odds
}
\description{
Convert Probabilities to Decimal odds
}
\examples{
odds.prob2dec(c(0.5,0.6))
}

