% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/odds.prob2hk.R
\name{odds.prob2hk}
\alias{odds.prob2hk}
\title{Convert Probabilities to Hong Kong odds}
\usage{
odds.prob2hk(x)
}
\arguments{
\item{x}{A vector of Probabilities}
}
\value{
A vector of Hong Kong odds
}
\description{
Convert Probabilities to Hong Kong odds
}
\examples{
odds.prob2hk(c(0.5,0.6))
}

