% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/odds.prob2indo.R
\name{odds.prob2indo}
\alias{odds.prob2indo}
\title{Convert Probabilities to Indonesian odds}
\usage{
odds.prob2indo(x)
}
\arguments{
\item{x}{A vector of Probabilities}
}
\value{
A vector of Indonesian odds
}
\description{
Convert Probabilities to Indonesian odds
}
\examples{
odds.prob2indo(c(0.5,0.6))
}

