% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/odds.prob2malay.R
\name{odds.prob2malay}
\alias{odds.prob2malay}
\title{Convert Probabilities to Malaysian odds}
\usage{
odds.prob2malay(x)
}
\arguments{
\item{x}{A vector of Probabilities}
}
\value{
A vector of Malaysian odds
}
\description{
Convert Probabilities to Malaysian odds
}
\examples{
odds.prob2malay(c(0.5,0.6))
}

