% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.prob2us.R
\name{odds.prob2us}
\alias{odds.prob2us}
\title{Convert Probabilities to US odds}
\usage{
odds.prob2us(x)
}
\arguments{
\item{x}{A vector of Probabilities}
}
\value{
A vector of US odds
}
\description{
Convert Probabilities to US odds
}
\examples{
odds.prob2us(c(0.5,0.6))
}
