% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/odds.2all.R
\name{odds.us2all}
\alias{odds.us2all}
\title{Convert US Odds to all other formats}
\usage{
odds.us2all(x)
}
\arguments{
\item{x}{A vector of US odds}
}
\value{
A dataframe
}
\description{
Convert US Odds to all other formats
}
\examples{
odds.us2all(c(-200,105))
}

