% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.us2dec.R
\name{odds.us2dec}
\alias{odds.us2dec}
\title{Convert US odds to Decimal odds}
\usage{
odds.us2dec(x)
}
\arguments{
\item{x}{A vector of US odds}
}
\value{
A vector of Decimal odds
}
\description{
Convert US odds to Decimal odds
}
\examples{
odds.us2dec(c(-200,150))
}
