% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.us2hk.R
\name{odds.us2hk}
\alias{odds.us2hk}
\title{Convert US odds to Hong Kong odds}
\usage{
odds.us2hk(x)
}
\arguments{
\item{x}{A vector of US odds}
}
\value{
A vector of Hong Kong odds
}
\description{
Convert US odds to Hong Kong odds
}
\examples{
odds.us2hk(c(-200,150))
}

