% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/odds.us2indo.R
\name{odds.us2indo}
\alias{odds.us2indo}
\title{Convert US odds to Indonesian odds}
\usage{
odds.us2indo(x)
}
\arguments{
\item{x}{A vector of US odds}
}
\value{
A vector of Indonesian odds
}
\description{
Convert US odds to Indonesian odds
}
\examples{
odds.us2indo(c(-200,150))
}

