% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.us2malay.R
\name{odds.us2malay}
\alias{odds.us2malay}
\title{Convert US odds to Malaysian odds}
\usage{
odds.us2malay(x)
}
\arguments{
\item{x}{A vector of US odds}
}
\value{
A vector of Malaysian odds
}
\description{
Convert US odds to Malaysian odds
}
\examples{
odds.us2malay(c(-200,150))
}

