% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.us2prob.R
\name{odds.us2prob}
\alias{odds.us2prob}
\title{Convert US odds to Probabilities}
\usage{
odds.us2prob(x)
}
\arguments{
\item{x}{A vector of US odds}
}
\value{
A vector of Probabilities
}
\description{
Convert US odds to Probabilities
}
\examples{
odds.us2prob(c(-200,150))
}
