% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.vig.R
\name{odds.vig}
\alias{odds.vig}
\title{Calculate the vigorish from a vector of vigged odds.}
\usage{
odds.vig(..., input = "us")
}
\arguments{
\item{...}{Odds of the format defined in input, the odds can be submitted as a data.frame or as a vector}

\item{input}{Default to US Odds. Valid selections are  "us" for US odds,
"prob" for Probability ,"dec" for Decimal odds,"malay" for Malaysian odds,
"hk" for Hong Kong odds, "indo" for Indonesian odds}
}
\value{
A vector of the vigs of the the given vigged odds
}
\description{
Calculate the vigorish from a vector of vigged odds.
}
\examples{
odds.vig(-120,110)
odds.vig(2.05,3.12,2.90,input="dec")
df <- data.frame(Home = c(1.5,1.8,1.9),
                Away = c(2.9,2.2,2.05))
odds.vig(df,input = "dec")
odds.vig(df$Home,df$Away,input="dec")
}
